refactor: reorganized interfaces and types in models.ts
- also little clean up
