+++ angular-frontend/src/app/device-handling/data-service/data.service.ts
@@ -1,7 +1,7 @@
 import { HttpClient } from '@angular/common/http';
 import { computed, inject, Injectable, signal } from '@angular/core';
-import { DataFormat, ServerDescription } from './omnAIServer';
-
+import { ServerDescription } from './omnAIServer';
+import { DataFormat } from './data.models';
 
 @Injectable({ providedIn: 'root' })
 export class DataService {
@@ -79,5 +79,4 @@ export class DataService {
     const server = this.servers()[serverURL];
     if (server) server.disconnect();
   }
-
 }
