+++ angular-frontend/src/app/device-handling/data-service/message.typeguards.ts
@@ -1,15 +1,9 @@
+import { omnAIScopeData } from './data.models';
 /**
  * Defines the possible WebSocket message formats.
  */
 export type wsMessageFormats = string | omnAIScopeData;
 
-interface omnAIScopeData {
-  devices: string[];
-  data: {
-    timestamp: number;
-    value: number[];
-  }[];
-}
 function isStringMessage(message: unknown): message is string {
   return typeof message === 'string';
 }
@@ -44,7 +38,7 @@ function isOmnAIDataMessage(message: unknown): message is omnAIScopeData {
 
 /**
  * Type guards for WebSocket message validation.
- * The websocket data are generated by the omnAI backend proces. Messages depend on the state.
+ * The websocket data are generated by the omnAI backend process. Messages depend on the state.
  * First, a message indicating a successful connection is send.
  * Then, omnAIscope data is captured and send in bulk.
  *
@@ -58,18 +52,3 @@ function isOmnAIDataMessage(message: unknown): message is omnAIScopeData {
  * {@link https://app.clickup.com/9015468401/v/dc/8cnu8bh-4415/8cnu8bh-5655}
  */
 export const messageTypeguards = { isOmnAIDataMessage, isStringMessage };
-
-/**
- * Defines the structure of a device overview response.
- *
- * The colors are RGB encoding for each device. The i-th device UUID has an LED
- * whose current color is encoded in the i-th entry of the colors array
- * */
-export interface DeviceOverview {
-  devices: {
-    UUID: string;
-  }[];
-  colors: {
-    color: { r: number; g: number; b: number };
-  }[];
-}
